+++ electron/main.ts
@@ -4,6 +4,7 @@ import { initDatabase, closeDatabase } from './database/connection'
 import { createTables } from './database/init'
 import { getOrCreateMasterKey } from './utils/keyManager'
 import { generateCaptcha, login, isFirstRun, initAdmin } from './services/auth'
+import { setDeviceMasterKey, listDevices, createDevice, updateDevice, deleteDevice, getDeviceById } from './services/device'
 
 let mainWindow: BrowserWindow | null = null
 let masterKey: string
@@ -32,6 +33,7 @@ function createWindow() {
 
 app.whenReady().then(() => {
   masterKey = getOrCreateMasterKey()
+  setDeviceMasterKey(masterKey)
   initDatabase()
   createTables()
 
@@ -40,6 +42,13 @@ app.whenReady().then(() => {
   ipcMain.handle('auth:login', (_e, u, p, ck, ci) => login(u, p, ck, ci))
   ipcMain.handle('auth:isFirstRun', () => isFirstRun())
   ipcMain.handle('auth:initAdmin', (_e, u, p) => initAdmin(u, p))
+
+  // Device IPC
+  ipcMain.handle('device:list', () => listDevices())
+  ipcMain.handle('device:create', (_e, data) => createDevice(data))
+  ipcMain.handle('device:update', (_e, id, data) => updateDevice(id, data))
+  ipcMain.handle('device:delete', (_e, id) => deleteDevice(id))
+  ipcMain.handle('device:getById', (_e, id) => getDeviceById(id))
 
   createWindow()
   app.on('activate', () => { if (BrowserWindow.getAllWindows().length === 0) createWindow() })